moves e2e tests to separate folder in e2e-tests folder
