+++ frontend/e2e-tests/tests/data-entry.e2e.ts
@@ -12,14 +12,14 @@ import {
 
 import { VotersCounts, VotesCounts } from "@kiesraad/api";
 
-import { test } from "./fixtures";
-import { fillDataEntry } from "./helpers-utils/e2e-test-helpers";
-import { formatNumber } from "./helpers-utils/e2e-test-utils";
+import { test } from "../fixtures";
+import { fillDataEntry } from "../helpers-utils/e2e-test-helpers";
+import { formatNumber } from "../helpers-utils/e2e-test-utils";
 import {
   noErrorsWarningsResponse,
   noRecountNoDifferencesDataEntry,
   noRecountNoDifferencesRequest,
-} from "./test-data/request-response-templates";
+} from "../test-data/request-response-templates";
 
 test.use({
   storageState: "e2e-tests/state/typist.json",
