+++ frontend/e2e-tests/tests/users.e2e.ts
@@ -1,13 +1,13 @@
 import { expect } from "@playwright/test";
 
-import { test } from "./fixtures";
-import { createRandomUsername } from "./helpers-utils/e2e-test-utils";
-import { UserCreateDetailsPgObj } from "./page-objects/users/UserCreateDetailsPgObj";
-import { UserCreateRolePgObj } from "./page-objects/users/UserCreateRolePgObj";
-import { UserCreateTypePgObj } from "./page-objects/users/UserCreateTypePgObj";
-import { UserDeleteModal } from "./page-objects/users/UserDeleteModalPgObj";
-import { UserListPgObj } from "./page-objects/users/UserListPgObj";
-import { UserUpdatePgObj } from "./page-objects/users/UserUpdatePgObj";
+import { test } from "../fixtures";
+import { createRandomUsername } from "../helpers-utils/e2e-test-utils";
+import { UserCreateDetailsPgObj } from "../page-objects/users/UserCreateDetailsPgObj";
+import { UserCreateRolePgObj } from "../page-objects/users/UserCreateRolePgObj";
+import { UserCreateTypePgObj } from "../page-objects/users/UserCreateTypePgObj";
+import { UserDeleteModal } from "../page-objects/users/UserDeleteModalPgObj";
+import { UserListPgObj } from "../page-objects/users/UserListPgObj";
+import { UserUpdatePgObj } from "../page-objects/users/UserUpdatePgObj";
 
 test.use({
   storageState: "e2e-tests/state/admin.json",
